Spike cross-correlation histogram
Altered function and parameter names of core CCH function

Clarified in documentation which spike train is lagging and which is
leading for CCH calculation.

Use new binning function in spikecorr.cch().

Removed old code from cch() function that is not needed with
BinnedSpikeTrain.

Simplified the way in which the smoothing kernel is supplied to the
cch() function.

Only arrays are allowed as kernels, the option to supply a string
leading to a function call to a specific numpy function creating a
kernel within was  removed.

Added TODOs for the future to cch() function header.

Long function name is better than cch

Changed the output to the cch to a neo.AnalogSignalArray and adjusted the
relative unit_test

Fixed bug in the trasformation of the histogram to AnalogSignalArray, before
the axis of the signal were inverted and following corrections to the unit
test

Fixed bug in the border_correction

Moved CCH function from to correct elephant module.

Added TODO where it is unclear how to compute the default window size.

Latest additions from code sprint.

Removed wrong auto-import

Fixed bug of the time axis of the cch, now the center bin is zero and the axis is symmetric respect to the center bin, udjusted also the check for the time axis check in the unit_test that now does not fail

Changed the definition of the histogram length to the length of the longest spike trains. Included two more tests, one for different bin sizes and one for when a window size is given.

Fixed bug where normalization and window options did not work together
(use of Hlen instead of Hbins)

3: Cross-correlation Histogram

Task-Url: http://github.com/INM-6/elephant/issues/issue/3
Signed-off-by: Michael Denker <m.denker@fz-juelich.de>

Put all data items of cch unit test into setup

Full coverage achieved with the test functions. Ready for pull request.

3: Cross-correlation Histogram

Task-Url: http://github.com/INM-6/elephant/issues/issue/3

Changed definition of time axis in CCH to be left bins. Smaller fixes in
documentation throughout.

Added an example for the CCH to the documentation.
3: Cross-correlation Histogram

Task-Url: http://github.com/INM-6/elephant/issues/issue/3

Changed all cch() to full name in unit tests
3: Cross-correlation Histogram

Task-Url: http://github.com/INM-6/elephant/issues/issue/3

Small changes to make unit tests more legible
3: Cross-correlation Histogram

Task-Url: http://github.com/INM-6/elephant/issues/issue/3

Change in error message